requirements.txt
amqp==2.6.1
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==4.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==2.2
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-formtools==2.3
django-import-export==2.8.0
django-redis==5.2.0
django-reversion==5.0.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
future==0.18.2
httplib2==0.9.2
idna==3.3
importlib-metadata==4.11.3
kombu==4.6.11
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
Pillow==9.1.0
prompt-toolkit==3.0.29
pyasn1==0.4.8
pycparser==2.21
pycrypto==2.6.1
pycryptodomex==3.9.4
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==19.1.0
pyparsing==3.0.8
python-alipay-sdk==3.0.4
pytz==2022.1
PyYAML==6.0
qcloudsms==0.21
qcloudsms-py==0.1.4
redis==4.2.2
requests==2.27.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
typing-extensions==4.1.1
urllib3==1.26.9
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
xadmin==0.6.1
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.0
